Fix parent parsing (for slices etc)
We want to return the parent type that knows the field, but when the parent is a slice (or a map, or its parent is another slice) the current approach did not work.
With this change, we walk up the tree until a non-slice, non-map type is found, so we can look up the struct field comment for it.